convenience typedefs, optimized map draw routine

+++ include/ui.h
@@ -2,43 +2,44 @@
 #define UI_H
 
 #include <stdbool.h>
-#include <stddef.h>
+
+#include "util.h"
 
 struct UiButton
 {
 	void (*Callback)(void);
 	char const *Text;
-	int x, y;
-	int w, h;
+	i32 x, y;
+	i32 w, h;
 	bool Hovered, Pressed;
 };
 
 struct UiTextField
 {
 	char *Out;
-	size_t Csr, Len, Nmax;
-	size_t FirstDraw, Ndraw;
-	int x, y;
-	int w, h;
+	u32 Csr, Len, Nmax;
+	u32 FirstDraw, Ndraw;
+	i32 x, y;
+	i32 w, h;
 	bool Hovered, Selected;
 };
 
 struct UiSlider
 {
 	void (*Callback)(float);
-	int x, y;
-	int w, h;
-	float Val;
+	i32 x, y;
+	i32 w, h;
+	f32 Val;
 	bool Hovered, Pressed;
 };
 
-struct UiButton UiButton_Create(int x, int y, char const *Text, void (*Callback)(void));
+struct UiButton UiButton_Create(i32 x, i32 y, char const *Text, void (*Callback)(void));
 void UiButton_Update(struct UiButton *Btn);
 void UiButton_Draw(struct UiButton const *Btn);
-struct UiTextField UiTextField_Create(int x, int y, size_t Ndraw, char *Out, size_t Nmax);
+struct UiTextField UiTextField_Create(i32 x, i32 y, u32 Ndraw, char *Out, u32 Nmax);
 void UiTextField_Update(struct UiTextField *Tf);
 void UiTextField_Draw(struct UiTextField const *Tf);
-struct UiSlider UiSlider_Create(int x, int y, int w, int h, float Initial, void (*Callback)(float));
+struct UiSlider UiSlider_Create(i32 x, i32 y, i32 w, i32 h, f32 Initial, void (*Callback)(f32));
 void UiSlider_Update(struct UiSlider *s);
 void UiSlider_Draw(struct UiSlider const *s);
